refactor: rename chunkmanager module to fetch
