+++ core/src/main/java/io/aiven/kafka/tieredstorage/fetch/ChunkKey.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.aiven.kafka.tieredstorage.chunkmanager;
+package io.aiven.kafka.tieredstorage.fetch;
 
 import java.nio.file.Path;
 import java.util.Objects;
