+++ core/src/main/java/io/aiven/kafka/tieredstorage/fetch/DefaultChunkManager.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.aiven.kafka.tieredstorage.chunkmanager;
+package io.aiven.kafka.tieredstorage.fetch;
 
 import java.io.InputStream;
 import java.util.List;
