+++ core/src/main/java/io/aiven/kafka/tieredstorage/fetch/cache/DiskBasedChunkCacheConfig.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.aiven.kafka.tieredstorage.chunkmanager.cache;
+package io.aiven.kafka.tieredstorage.fetch.cache;
 
 import java.io.IOException;
 import java.nio.file.Files;
