+++ core/src/test/java/io/aiven/kafka/tieredstorage/fetch/DefaultChunkManagerTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.aiven.kafka.tieredstorage.chunkmanager;
+package io.aiven.kafka.tieredstorage.fetch;
 
 import javax.crypto.Cipher;
